🎨 Enforce ordered imports (#197)
I'd like to keep our project consistent by ordering Javascript imports. This PR turns on the `eslint` rule to enforce consistent ordering. To help out with this, I've added a VSCode extension to the projects extension recommendations that does this automatically on save. Install the recommended extensions and this should be pretty painless.