:bug: [server] Disable service worker caching for auth routes
Sapper includes a service worker that caches pages. I know this is
causing problems with logging out, and I suspect it may be causing the
log in problems as well. Test for those routes in the service worker
logic, and if we're on one of those routes handle the request in the
traditional way.

Bonus: add some detail to grant errors where id_token is not defined.